refactor: 폴더 구조 리팩토링 및 ManageController 클래스 선언부분 인터페이스로 이동
