--- src/main/java/subway/controller/Controller.java
@@ -1,9 +1,0 @@
-package subway.controller;
-
-import subway.config.constants.views.Targets;
-
-public interface Controller {
-    public boolean register();
-    public boolean delete();
-    public boolean read();
-}
