+++ src/main/java/subway/controller/ManageController.java
@@ -1,8 +1,7 @@
 package subway.controller;
 
-import subway.view.AskView;
-import subway.view.ResponseView;
-import subway.view.util.MakeString;
+import subway.controller.utils.Constants;
+import subway.controller.utils.Controller;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
@@ -10,10 +9,6 @@ import java.io.InputStreamReader;
 import static subway.controller.StationController.subwayException;
 
 public abstract class ManageController implements Controller, Constants {
-    static final AskView ask = new AskView();
-    static final ResponseView response = new ResponseView();
-    static final MakeString makeString = new MakeString();
-
     protected static BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
     public boolean work(final Controller controller, final String target) {
@@ -21,8 +16,7 @@ public abstract class ManageController implements Controller, Constants {
         ask.Function();
         try {
             String command = br.readLine();
-            sendRequest(controller, command);
-            return true;
+            return sendRequest(controller, command);
         } catch (Exception e) {
             subwayException.unexpected();
         }
