+++ src/main/java/subway/controller/utils/Methods.java
@@ -1,0 +1,6 @@
+package subway.controller.utils;
+
+public class Methods {
+    public Methods(){}
+
+}
